+++ cmd/microservice/git-test.go
@@ -11,6 +11,7 @@ import (
 	gitStorage "github.com/dolittle/platform-api/pkg/platform/storage/git"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
 var gitTestCMD = &cobra.Command{
@@ -21,7 +22,8 @@ var gitTestCMD = &cobra.Command{
 		logrus.SetOutput(os.Stdout)
 
 		logContext := logrus.StandardLogger()
-		gitRepoConfig := git.InitGit(logContext)
+		platformEnvironment := viper.GetString("tools.server.platformEnvironment")
+		gitRepoConfig := git.InitGit(logContext, platformEnvironment)
 
 		gitRepo := gitStorage.NewGitStorage(
 			logrus.WithField("context", "git-repo"),